Login/requirements.txt
Flask>=2.1.2
flask-cors==3.0.10
flask-httpauth==4.8.0
Flask-SQLAlchemy
Flask-JWT-Extended
SQLAlchemy==2.0.20
python-dotenv==1.0.0
PyJWT==2.8.0
pathspec==0.11.2
Flask-Migrate==3.1.0